BIGUPDATE

+++ srcs/string/ft_memcpy.c
@@ -6,7 +6,7 @@
 /*   By: mkoyamba <mkoyamba@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/10/05 14:59:21 by mkoyamba          #+#    #+#             */
-/*   Updated: 2021/10/05 15:47:55 by mkoyamba         ###   ########.fr       */
+/*   Updated: 2021/10/09 10:37:47 by mkoyamba         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,7 +16,7 @@ void	*ft_memcpy(void *restrict dst, const void *restrict src, size_t n)
 {
 	char	*tdst;
 	char	*tsrc;
-	int		a;
+	size_t	a;
 
 	tdst = (char *)dst;
 	tsrc = (char *)src;
